Assignment Completed

+++ src/nuber/students/BookingResult.java
@@ -2,12 +2,12 @@ package nuber.students;
 
 public class BookingResult {
 
-	public int jobID;
+	public long jobID;
 	public Passenger passenger;
 	public Driver driver;
 	public long tripDuration;
 	
-	public BookingResult(int jobID, Passenger passenger, Driver driver, long tripDuration)
+	public BookingResult(long jobID, Passenger passenger, Driver driver, long tripDuration)
 	{
 		this.jobID = jobID;
 		this.passenger = passenger;
